move table state out as it is not relevent to optimize table
