+++ src/components/gui/table-result/context-menu.tsx
@@ -11,7 +11,7 @@ import { LucidePlus, LucideTrash2 } from "lucide-react";
 import { useCallback } from "react";
 import { useFullEditor } from "../providers/full-editor-provider";
 import OptimizeTableState from "../table-optimized/optimize-table-state";
-import TableStateActions from "../table-optimized/table-state-actions";
+import TableStateActions from "./table-state-actions";
 import { TableHeaderMetadata } from "./type";
 
 export default function useTableResultContextMenu({
